+++ sdk/core/azure_core/benches/http_transport_benchmarks.rs
@@ -5,7 +5,7 @@ use azure_core::{
     credentials::TokenCredential,
     fmt::SafeDebug,
     http::{
-        ClientMethodOptions, ClientOptions, HttpClient, Method, Pipeline, RawResponse, Request,
+        BufResponse, ClientMethodOptions, ClientOptions, HttpClient, Method, Pipeline, Request,
         TransportOptions, Url,
     },
     Result,
@@ -85,7 +85,7 @@ impl TestServiceClient {
         &self,
         path: &str,
         options: Option<TestServiceClientGetMethodOptions<'_>>,
-    ) -> Result<RawResponse> {
+    ) -> Result<BufResponse> {
         let options = options.unwrap_or_default();
         let mut url = self.endpoint.clone();
         url.set_path(path);
@@ -103,6 +103,7 @@ impl TestServiceClient {
                 azure_core::error::ErrorKind::HttpResponse {
                     status: response.status(),
                     error_code: None,
+                    raw_response: None,
                 },
                 format!("Failed to GET {}: {}", request.url(), response.status()),
             ));