+++ sdk/core/azure_core_macros/src/tracing_function.rs
@@ -498,11 +498,11 @@ mod tests {
                 let ctx = options.method_options.context.clone();
                 let pipeline = pipeline.clone();
                 async move {
-                    let rsp: RawResponse = pipeline.send(&ctx, &mut request).await?;
+                    let rsp: BufResponse = pipeline.send(&ctx, &mut request).await?;
                     let (status, headers, body) = rsp.deconstruct();
                     let bytes = body.collect().await?;
                     let res: ListDeletedSecretPropertiesResult = json::from_json(&bytes)?;
-                    let rsp = RawResponse::from_bytes(status, headers, bytes).into();
+                    let rsp = BufResponse::from_bytes(status, headers, bytes).into();
                     Ok(match res.next_link {
                         Some(next_link) if !next_link.is_empty() => PagerResult::More {
                             response: rsp,
@@ -569,11 +569,11 @@ mod tests {
                     let ctx = options.method_options.context.clone();
                     let pipeline = pipeline.clone();
                     async move {
-                        let rsp: RawResponse = pipeline.send(&ctx, &mut request).await?;
+                        let rsp: BufResponse = pipeline.send(&ctx, &mut request).await?;
                         let (status, headers, body) = rsp.deconstruct();
                         let bytes = body.collect().await?;
                         let res: ListDeletedSecretPropertiesResult = json::from_json(&bytes)?;
-                        let rsp = RawResponse::from_bytes(status, headers, bytes).into();
+                        let rsp = BufResponse::from_bytes(status, headers, bytes).into();
                         Ok(match res.next_link {
                             Some(next_link) if !next_link.is_empty() => PagerResult::More {
                                 response: rsp,