+++ sdk/typespec/typespec_client_core/src/http/clients/reqwest.rs
@@ -5,7 +5,7 @@ use crate::http::{
     headers::{HeaderName, HeaderValue, Headers},
     request::{Body, Request},
     response::PinnedStream,
-    HttpClient, Method, RawResponse, Sanitizer, DEFAULT_ALLOWED_QUERY_PARAMETERS,
+    BufResponse, HttpClient, Method, Sanitizer, DEFAULT_ALLOWED_QUERY_PARAMETERS,
 };
 use async_trait::async_trait;
 use futures::TryStreamExt;
@@ -33,7 +33,7 @@ pub fn new_reqwest_client() -> Arc<dyn HttpClient> {
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 impl HttpClient for ::reqwest::Client {
-    async fn execute_request(&self, request: &Request) -> Result<RawResponse> {
+    async fn execute_request(&self, request: &Request) -> Result<BufResponse> {
         let url = request.url().clone();
         let method = request.method();
         let mut req = self.request(from_method(method), url.clone());
@@ -74,7 +74,7 @@ impl HttpClient for ::reqwest::Client {
             )
         }));
 
-        Ok(RawResponse::new(status.as_u16().into(), headers, body))
+        Ok(BufResponse::new(status.as_u16().into(), headers, body))
     }
 }