+++ sdk/typespec/typespec_client_core/src/http/policies/transport.rs
@@ -71,7 +71,7 @@ impl Header for EmptyContentLength {
 #[cfg(all(test, not(target_family = "wasm")))]
 mod tests {
     use super::*;
-    use crate::http::{headers::Headers, RawResponse, StatusCode};
+    use crate::http::{headers::Headers, BufResponse, StatusCode};
 
     #[derive(Debug)]
     struct MockTransport;
@@ -84,7 +84,7 @@ mod tests {
             _request: &mut Request,
             _next: &[Arc<dyn Policy>],
         ) -> PolicyResult {
-            PolicyResult::Ok(RawResponse::from_bytes(
+            PolicyResult::Ok(BufResponse::from_bytes(
                 StatusCode::Ok,
                 Headers::new(),
                 Vec::new(),